Avoid redundant call to CGContextClipToRect in GraphicsContextCG::fillRect when filling with a gradient https://bugs.webkit.org/show_bug.cgi?id=276623 <radar://131632207>
Reviewed by NOBODY (OOPS!).

Skip CGContextClipToRect call when the rect is completely contains
the current clipping rect. The result of intersecting the clipping
rectangle with a rectangle which completely contains it is equivalent
to a nop.

* Source/WebCore/platform/graphics/cg/GraphicsContextCG.cpp:
(WebCore::GraphicsContextCG::fillRect):
Avoid save / restore cost in the no-clipping case, since we can
simply apply the inverse of the affine transform which is less
expensive than a save and restore.