REGRESSION (277477@main): Site persistence issues caused by passing incorrect span to SQLiteStatement::bindBlob https://bugs.webkit.org/show_bug.cgi?id=276791 rdar://131333449
Unreviewed, partial revert of 277477@main.

In 277477@main, we adopted `std::span` in SQLiteStatement::bindBlob.
That patch introduced a subtle change the span we were passing to the
bindBlob specalization that accepts a `std::span` argument. Instead of
passing an empty span, we started passing a span with a null character
in it. This lead to some downstream site persistence issues, such as
unresponsive navbars on weather.com.

We partially revert 277477@main (the changes to SQLiteStatement.cpp) to
restore our original behavior.

* Source/WebCore/platform/sql/SQLiteStatement.cpp:
(WebCore::SQLiteStatement::bindBlob):

Canonical link: https://commits.webkit.org/281097@main