Avoid crashing if traverseUnchangedSubtree can't find the same backingSharingProvider. https://bugs.webkit.org/show_bug.cgi?id=276754 <rdar://131972979>
Reviewed by Simon Fraser.

This currently is a nullptr crash if backingProviderCandidateForLayer returns
nullptr.

We should be able to see if the candidate used from the last update is still
available, and use that to avoid the crash (but still render incorrectly).

* Source/WebCore/rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::BackingSharingState::previousBackingProviderCandidateForLayer):
(WebCore::RenderLayerCompositor::traverseUnchangedSubtree):

Canonical link: https://commits.webkit.org/281112@main