http/wpt/mediastream/mediastreamtrackprocessor-videoframe-timestamp.html is flakey rdar://131979190 https://bugs.webkit.org/show_bug.cgi?id=276763
Reviewed by Eric Carlson.

The test is synchronously observing a VideoTrackGenerator via a MediaStreamTrackProcessor and enqueuing a VideoFrame in the VideoTrackGenerator.
Before the patch, the observer registration would require hopping to the main thread, which would race with the enqueuing of the VideoFrame.
We are now using callOnMainThreadAndWait to stop the worker until the observer is registered.
The enqueuing of the VideoFrame will happen after the registration and the test will be able to proceed without any issue.

* LayoutTests/TestExpectations:
* Source/WebCore/Modules/mediastream/MediaStreamTrackProcessor.cpp:
(WebCore::MediaStreamTrackProcessor::VideoFrameObserverWrapper::start):

Canonical link: https://commits.webkit.org/281125@main