Improve "congratulations" text in build-webkit by having the complete command to run needed a launcher https://bugs.webkit.org/show_bug.cgi?id=276892
Reviewed by Alexey Proskuryakov.

Makes it more clear what exactly to do after build-webkit has completed
by printing out the complete command to run in order to run the launcher.

Goes from:

====================================================================
 WebKit is now built (10m:31s).
 To run Safari with this newly-built code, use the
 "Tools/Scripts/run-safari" script.
====================================================================

to

====================================================================
 WebKit is now built (10m:31s).
 To run Safari with this newly-built code, use
 the command "Tools/Scripts/run-safari --debug".
====================================================================

This avoids the problem of trying to run the launcher and getting confusing
errors when the current configuration does not match what was built.

* Tools/Scripts/build-webkit:
(writeCongrats): Append argumentsForConfiguration() to the path to get
the full command to run for the current invocation of build-webkit.

Canonical link: https://commits.webkit.org/281246@main